moved poc into table storybook cases
